Refactor: separated type and implementation
